% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollin_for_list_cities.R
\name{get_pollen_for_list_cities}
\alias{get_pollen_for_list_cities}
\title{Get data pollen of a given city}
\usage{
get_pollen_for_list_cities(list_cities, PATH = "SwissCities.csv")
}
\arguments{
\item{list_cities}{list of cities in a string form}

\item{PATH}{path of the cities file}
}
\value{
A tibble with city name, canton, lat, lng, pollin level
}
\description{
Retrieve information about the pollen of given cities
}
\examples{
\dontrun{
tibble_test <- get_pollen_for_list_cities(c("Lugano", "Geneva"), "SwissCities.csv")
}

}
